feat: add a flat display when there are no values
